+++ Engine.h
@@ -76,9 +76,9 @@ public:
 
     irr::core::vector3df camTarget();
     void run();
-    void loadMesh(const std::wstring& fileName);
-    void reloadMesh();
-    void saveMesh(const irr::io::path path);
+    bool loadMesh(const std::wstring& fileName);
+    bool reloadMesh();
+    std::wstring saveMesh(const irr::io::path path, const std::string& nameOrBlank, const std::string& extension);
     void reloadTexture();
     bool loadTexture(const std::wstring& fileName);
     void setMeshDisplayMode(bool wireframe = false, bool lighting = true, bool textureInterpolation = true);